requirements.txt
# 🗣️ COMMUNICATION CLASSIFICATION SYSTEM
# Requirements untuk klasifikasi Verbal vs Non-Verbal dengan KNN & PSO

# Core Machine Learning
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Computer Vision & Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# Web Framework
Flask>=2.0.0
Flask-CORS>=3.0.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Scientific Computing
scipy>=1.7.0

# Excel Processing
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# API & HTTP
requests>=2.25.0

# Utilities
joblib>=1.0.0
tqdm>=4.60.0

# Optional: Advanced Deep Learning (untuk feature extraction tambahan)
# tensorflow>=2.6.0
# torch>=1.9.0
# torchvision>=0.10.0

# Development & Testing
pytest>=6.0.0